refactor: Remove _compat
