+++ libvcs/util.py
@@ -7,11 +7,23 @@ import errno
 import logging
 import os
 import subprocess
+import sys
 
 from . import exc
-from ._compat import console_to_str
 
 logger = logging.getLogger(__name__)
+
+console_encoding = sys.__stdout__.encoding
+
+
+def console_to_str(s):
+    """ From pypa/pip project, pip.backwardwardcompat. License MIT. """
+    try:
+        return s.decode(console_encoding)
+    except UnicodeDecodeError:
+        return s.decode('utf_8')
+    except AttributeError:  # for tests, #13
+        return s
 
 
 def which(
