+++ setup.cfg
@@ -1,5 +1,5 @@
 [flake8]
-exclude = .*/,.tox,*.egg,libvcs/_compat.py,libvcs/__*__.py,
+exclude = .*/,.tox,*.egg,libvcs/__*__.py,
 select = E,W,F,N
 max-line-length = 88
 # Stuff we ignore thanks to black: https://github.com/ambv/black/issues/429
